major revamp to components housekeep
