+++ src/app/_components/post_gen/hero-post.tsx
@@ -2,7 +2,7 @@ import Avatar from "@/app/_components/avatar";
 import CoverImage from "@/app/_components/cover-image";
 import { type Author } from "@/interfaces/author";
 import Link from "next/link";
-import DateFormatter from "./date-formatter";
+import DateFormatter from "../date-formatter";
 
 type Props = {
   title: string;
