+++ src/app/_components/post_gen/post-body-client.tsx
