+++ src/app/_components/post_gen/post-header.tsx
@@ -1,7 +1,7 @@
-import Avatar from "./avatar";
-import CoverImage from "./cover-image";
-import DateFormatter from "./date-formatter";
-import { PostTitle } from "@/app/_components/post-title";
+import Avatar from "../avatar";
+import CoverImage from "../cover-image";
+import DateFormatter from "../date-formatter";
+import { PostTitle } from "@/app/_components/post_gen/post-title";
 import { type Author } from "@/interfaces/author";
 
 type Props = {
