+++ src/app/_components/post_gen/post-preview.tsx
@@ -1,8 +1,8 @@
 import { type Author } from "@/interfaces/author";
 import Link from "next/link";
-import Avatar from "./avatar";
-import CoverImage from "./cover-image";
-import DateFormatter from "./date-formatter";
+import Avatar from "../avatar";
+import CoverImage from "../cover-image";
+import DateFormatter from "../date-formatter";
 
 type Props = {
   title: string;
