+++ src/app/topics/page.tsx
@@ -1,7 +1,7 @@
 import Container from "@/app/_components/container";
 import { getAllPosts } from "@/lib/api";
-import { HeroPost } from "@/app/_components/hero-post";
-import { MoreStories } from "@/app/_components/more-stories";
+import { HeroPost } from "@/app/_components/post_gen/hero-post";
+import { MoreStories } from "@/app/_components/post_gen/more-stories";
 
 export default function Index() {
   const allPosts = getAllPosts("_topics");
